Interim updates: coordinator page, now-racing.js

+++ website/coordinator.php
@@ -21,6 +21,10 @@ require_once('inc/timer-state.inc');
   margin: 10px;
   padding: 10px;
  }
+
+.current {
+    background-color: #F7D117;
+ }
 </style>
 </head>
 <body>
@@ -43,38 +47,21 @@ function scan_kiosks($prefix, $kiosk_page) {
 ?>
 <div class="control_column">
 
-<div class="kiosk_control_group">
+
+<div class="control_group heat_control_group">
+  <!-- TODO: Next heat/previous heat, manual heat results -->
+  <h3>Heat Control</h3>
+
 <div class="block_buttons">
-<?php
-try {
-    $stmt = $db->query('SELECT address, name, page, last_contact FROM Kiosks ORDER BY name, last_contact');
-    foreach ($stmt as $row) {
-        echo '<div class="control_group kiosk_control">'."\n";
-        echo   '<p>Kiosk ';
-        echo   '<span class="kiosk_control_name">'.htmlspecialchars($row['name']).'</span> ';
-        echo   '<span class="kiosk_control_address">'.htmlspecialchars($row['address']).'</span> ';
-        // TODO: last_contact
-        echo   '</p>';
 
-        echo   '<label for="kiosk-page">Display:</label>'; // TODO: Fire action when selection changes.
-        echo   '<select name="kiosk-page" data-kiosk-address="'.htmlspecialchars($row['address']).'">'."\n";
-        echo   '<optgroup>'."\n";
+<input type="button" value="Skip Heat"/><br/>
+<input type="button" value="Previous Heat"/><br/>
+<input type="button" value="Manual Heat Results"/><br/>
 
-        scan_kiosks('kiosks', $row['page']);
-        scan_kiosks('local'.DIRECTORY_SEPARATOR.'kiosks', $row['page']);
-
-        echo   '</optgroup>'."\n";
-        echo   '</select>'."\n";
-        echo '</div>'."\n";
-    }
-} catch (PDOException $p) {
-    // TODO: No Kiosks table
-    echo '<h2>No Kiosks table defined.</h2>';  // TODO: Or other problem...
-    // TODO: What if not throwing exceptions?
-}
-?>
 </div>
+
 </div>
+
 
 <div class="control_group timer_control_group">
   <!-- TODO: Show timer status -->
@@ -112,14 +99,45 @@ try {
   <h3>Instant Replay Status</h3>
 </div>
 
+
+
+<div class="kiosk_control_group">
+<div class="block_buttons">
+<?php
+try {
+    $stmt = $db->query('SELECT address, name, page, last_contact FROM Kiosks ORDER BY name, last_contact');
+    foreach ($stmt as $row) {
+        echo '<div class="control_group kiosk_control">'."\n";
+        echo   '<p>Kiosk ';
+        echo   '<span class="kiosk_control_name">'.htmlspecialchars($row['name']).'</span> ';
+        echo   '<span class="kiosk_control_address">'.htmlspecialchars($row['address']).'</span> ';
+        // TODO: last_contact
+        echo   '</p>';
+
+        echo   '<label for="kiosk-page">Display:</label>'; // TODO: Fire action when selection changes.
+        echo   '<select name="kiosk-page" data-kiosk-address="'.htmlspecialchars($row['address']).'">'."\n";
+        echo   '<optgroup>'."\n";
+
+        scan_kiosks('kiosks', $row['page']);
+        scan_kiosks('local'.DIRECTORY_SEPARATOR.'kiosks', $row['page']);
+
+        echo   '</optgroup>'."\n";
+        echo   '</select>'."\n";
+        echo '</div>'."\n";
+    }
+} catch (PDOException $p) {
+    // TODO: No Kiosks table
+    echo '<h2>No Kiosks table defined.</h2>';  // TODO: Or other problem...
+    // TODO: What if not throwing exceptions?
+    // TODO: Create kiosk table on demand
+    // TODO: Allow assigning name to kiosk
+}
+?>
+</div>
+</div>
 </div>
 
 <div class="control_column">
-
-<div class="control_group heat_control_group">
-  <!-- TODO: Next heat/previous heat, manual heat results -->
-  <h3>Heat Control</h3>
-</div>
 
 <div class="scheduling_control_group">
 <div class="block_buttons">
@@ -133,20 +151,64 @@ $curr_round = get_running_round();
 $stmt = $db->query('SELECT roundid, Classes.class, round FROM Rounds'
                    .' INNER JOIN Classes ON Rounds.classid = Classes.classid'
                    .' ORDER BY round, Classes.class');
+$rounds = array();
 foreach ($stmt as $round) {
-    echo '<div class="control_group scheduling_control">'."\n";
+    $rounds[] = $round;
+}
+
+foreach ($rounds as $round) {
+    $roundid = $round['roundid'];
+
+    // Schedule/reschedule: if exist roster members not in schedule.
+    // Race if a schedule exists and not presently racing.
+    // Discard if there are results.
+    $unscheduled = read_single_value('SELECT COUNT(*)'
+                                     .' FROM Roster'
+                                     .' INNER JOIN RegistrationInfo'
+                                     .' ON Roster.racerid = RegistrationInfo.racerid'
+                                     .' WHERE Roster.roundid = :roundid'
+                                     .' AND RegistrationInfo.passedinspection <> 0'
+                                     .' AND NOT EXISTS(SELECT 1 FROM RaceChart'
+                                     .'  WHERE RaceChart.roundid = Roster.roundid'
+                                     .'  AND RaceChart.racerid = Roster.racerid)',
+                                     array(':roundid' => $roundid));
+    $already_run = read_single_value('SELECT COUNT(*) FROM RaceChart'
+                                     .' WHERE roundid = :roundid'
+                                     .' AND finishtime IS NOT NULL',
+                                     array(':roundid' => $roundid));
+    echo '<div class="control_group scheduling_control'
+    .(@$curr_round['roundid'] == $round['roundid'] ? ' current' : '')
+    .'">'."\n";
     echo '<p>'.htmlspecialchars($round['class']).', round '.$round['round'].'</p>'."\n";
+
+    if ($unscheduled) {
+        echo '<p>'.$unscheduled.' unscheduled roster members</p>';
+        if ($already_run) {
+            echo '<input type="button" value="Reschedule"/>'."\n";
+        } else {
+            echo '<input type="button" value="Schedule"/>'."\n";
+        }
+    }
     // TODO: Actions for these buttons
     // TODO: Buttons enable/disable according to state of the round.
     // TODO: States for a round: Not scheduled; scheduled but not raced; partial results; completed
     // TODO: Count racers in a round/roster, count passed racers in a round
-    echo '<input type="button" value="Schedule"/>'."\n";
-    echo '<input type="button" value="Reschedule"/>'."\n";
-    echo '<input type="button" value="Race"/>'."\n";
-    echo '<input type="button" value="Discard Results"/>'."\n";
     if (@$curr_round['roundid'] == $round['roundid']) {
         // TODO: Master schedule?
         echo '<p>(Currently racing)</p>';
+    } else {
+        $scheduled_but_not_run = read_single_value('SELECT COUNT(*)'
+                                                   .' FROM RaceChart'
+                                                   .' WHERE roundid = :roundid'
+                                                   .' AND finishtime IS NULL',
+                                                   array(':roundid' => $roundid));
+        if ($scheduled_but_not_run) {
+            echo '<input type="button" value="Race"/>'."\n";
+        }
+    }
+
+    if ($already_run) {
+        echo '<input type="button" value="Discard Results"/>'."\n";
     }
     echo '</div>'."\n";
 }
